+++ src/app/admin/components/certifiedProduct/management.directive.spec.js
@@ -103,7 +103,7 @@
                 commonService.getSearchOptions.and.returnValue($q.when({}));
                 commonService.getTargetedUsers.and.returnValue($q.when([]));
                 commonService.getTestFunctionality.and.returnValue($q.when({data: []}));
-                commonService.getTestStandards.and.returnValue($q.when([]));
+                commonService.getTestStandards.and.returnValue($q.when({data: []}));
                 commonService.getTestTools.and.returnValue($q.when([]));
                 commonService.getUcdProcesses.and.returnValue($q.when([]));
                 commonService.getUploadingCps.and.returnValue($q.when(mock.uploadingCps));